Algorithmic_Strategies/Rates_Carry/requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
yfinance>=0.2.28
pandas-datareader>=0.10.0
fredapi>=0.5.0
statsmodels>=0.14.0
scikit-learn>=1.3.0
PyYAML>=6.0
jupyter>=1.0.0
notebook>=7.0.0
